% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_corr.R
\name{calculate_corr}
\alias{calculate_corr}
\title{Calculate correlation between variables in the dataset}
\usage{
calculate_corr(var)
}
\arguments{
\item{var}{name of variable}
}
\value{
A dataframe with correlation and p-value
}
\description{
This function takes 1  variable from \code{wildlife_strikes} dataset and computes the correlation and p-value between that selected variable and repair costs incurred on airline due to wildlife strikes.
}
\examples{
calculate_corr(var = height)
calculate_corr(var = speed)
}
