% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_airline.R
\name{cases_airline}
\alias{cases_airline}
\title{Return the number of attacks incurred by selected airline in the selected year}
\usage{
cases_airline(input_operator, input_year)
}
\arguments{
\item{input_operator}{Name of the airlines in the dataset, namely \verb{AMERICAN AIRLINES}, \verb{UNITED AIRLINES}, \verb{SOUTHWEST AIRLINES} and \verb{DELTA AIR LINES}.}

\item{input_year}{Year in the dataset, ranging from 1991 to 2018}
}
\value{
Characters
}
\description{
This function calculates the number of wildlife attacks incurred by selected airline in the selected year. This function is also being used in Shiny App.
}
\examples{
cases_airline("AMERICAN AIRLINES", 2018)
cases_airline("UNITED AIRLINES", 2018)


}
