% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_state.R
\name{cases_statemap}
\alias{cases_statemap}
\alias{cases_state}
\title{Compute number of cases of wildlife strike reported in each state}
\usage{
cases_statemap(input_state)

cases_state(input_state)
}
\arguments{
\item{input_state}{Name of the states in the dataset, such as california, alabama, flordia and georgia.}
}
\value{
A choropleth map showing the reported number of wildlife strikes per state when \code{cases_statemap(input_state)} is run or a dataframe  when \code{cases_state(input_state)} is run.
}
\description{
These functions calculate the number of wildlife strikes reported in each state from 1991 to 2018 in the \code{wildlife_strikes} dataset.
Run \code{cases_state(input_state)} to obtain the observations for the selected states in a tibble format. Run \code{cases_statemap(input_state)} which creates a choropleth map showing the number of cases in the selected states. This function is also being used in Shiny App.
}
\examples{
cases_state(c("california", "alabama"))
cases_state(c("arizona", "california", "florida", "colorado", "texas"))
cases_statemap(c("california", "alabama"))
cases_statemap(c("arizona", "california", "florida", "colorado", "texas"))

}
