% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_frequency.R
\name{count_frequency}
\alias{count_frequency}
\alias{count_frequency_plot}
\title{Calculate frequency of occurrence}
\usage{
count_frequency(var)

count_frequency_plot(var)
}
\arguments{
\item{var}{name of variable}
}
\value{
A dataframe with observations and frequency, similar to the function \code{df \%>\% summarise (n = n())} when \code{count_frequency(var)} is run or bar plot when \code{count_frequency_plot(var)} is run.
}
\description{
These functions calculate the frequency of the observations for a particular variable in the \code{wildlife_strikes} dataset.
It is helpful in identifying which condition in the variable is associated with higher occurence of wildlife strikes to aircrafts.
Run \code{count_frequency(var)} to obtain the frequency of observations in a tibble format. Run \code{count_frequency_plot(var)} which shows the number of cases in each group in a bar plot format.
This function is also being used in Shiny App.
}
\examples{
count_frequency(sky)
count_frequency(phase_of_flt)
count_frequency_plot(sky)
count_frequency_plot(phase_of_flt)

}
