% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damages_airline.R
\name{damages_airline}
\alias{damages_airline}
\title{Computes the number of attacks incurred by selected airline in the selected year  by damages level}
\usage{
damages_airline(input_operator, input_year)
}
\arguments{
\item{input_operator}{Name of the airlines in the dataset, namely \verb{AMERICAN AIRLINES}, \verb{UNITED AIRLINES}, \verb{SOUTHWEST AIRLINES} and \verb{DELTA AIR LINES}}

\item{input_year}{Year in the dataset, ranging from 1990 to 2018}
}
\value{
A dataframe.
}
\description{
This function computes the number of wildlife attacks incurred by airlines by damages level. This function is also being used in Shiny App.
}
\examples{
damages_airline("AMERICAN AIRLINES", 2011)
damages_airline("UNITED AIRLINES", 2011)

}
