% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{histPlot}
\alias{histPlot}
\title{Histogram for repair cost for the particular airline in the year}
\usage{
histPlot(input_operator, input_bins)
}
\arguments{
\item{input_operator}{Name of the airlines in the dataset, namely \verb{AMERICAN AIRLINES}, \verb{UNITED AIRLINES}, \verb{SOUTHWEST AIRLINES} and \verb{DELTA AIR LINES}}

\item{input_bins}{Number of the bins for histogram}
}
\value{
A histogram plot which displays the distribution of with selected number of bins.
}
\description{
This function builds the histogram which shows the distribution of repair cost incurred for the selected airline from 1991 to 2018 due to wildlife strikes, where data is obtained from the \code{wildlife_strikes} dataset. This function will be used in the Shiny App.
}
\examples{
histPlot("AMERICAN AIRLINES", 10)
histPlot("UNITED AIRLINES", 5)
}
