% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Load Shiny Application from inst/app directory}
\usage{
run_app()
}
\value{
A Shiny App embedded in the package
}
\description{
Load Shiny Application from inst/app directory
}
