% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_plot.R
\name{theme_plot}
\alias{theme_plot}
\title{Create custom theme for plots}
\usage{
theme_plot(plot_type)
}
\arguments{
\item{plot_type}{The type of plot - "map" or "bar"}
}
\value{
Custom theme that complements the aesthetics of the Shiny App
}
\description{
Create custom theme for bar charts, maps and histogram.
}
