% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_inputs.R
\name{ui_input}
\alias{ui_input}
\title{Creates user interface inputs for Shiny app}
\usage{
ui_input(input_type, input_id, dataset)
}
\arguments{
\item{input_type}{Character string specifying type of input, "date" or "state"}

\item{input_id}{A string specifying the label Id}

\item{dataset}{A data frame}
}
\value{
Creates a reactive slider/picker input for Shiny App
}
\description{
This function creates user interface inputs for Shinyapp such as radio button and slider input.
}
