% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USmap.R
\docType{data}
\name{USmap}
\alias{USmap}
\title{Raw US map data for states}
\format{
a tibble with 15527 observations and 5 variables
\itemize{
\item \strong{region}: U.S state name
\item \strong{abb}: Two-letter U.S state code
\item \strong{long}: Longitude of the U.S state
\item \strong{lat}: Latitude of the U.S state
\item \strong{group}: Geographical coordinates for a group of observations/states
}
}
\source{
ggplot 2 R package which turns data from the maps package in to a data frame suitable for plotting.
}
\usage{
USmap
}
\description{
Raw US map data for states
}
\keyword{datasets}
