% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildlife_strikes.R
\docType{data}
\name{wildlife_strikes}
\alias{wildlife_strikes}
\title{Provides the Dataset for reported wildlife strikes from 1991 to 2018 in the U.S, reported by The New York Times.}
\format{
a tibble with 225 observations and 21 variables:
\itemize{
\item \strong{incident_date}: Date of incident
\item \strong{abb}: Abbreviation of U.S. state
\item \strong{airport_id}: Airport ID
\item \strong{airport}: Airport Name
\item \strong{operator}: Operator/Airline
\item \strong{atype}: Airline type
\item \strong{type_eng}: Engine type
\item \strong{species_id}: Species ID
\item \strong{species}: Species
\item \strong{damage}: Damage: N None M Minor, M Uncertain, S Substantial, D Destroyed
\item \strong{num_engs}: Number of engines
\item \strong{incident_month}: Incident month
\item \strong{incident_year}: Incident year
\item \strong{time_of_day}: Incident time
\item \strong{time}: Incident time
\item \strong{height}: Plane height at impact
\item \strong{speed}: Plane speed at impact
\item \strong{phase_of_flt}: Phase of flight at impact
\item \strong{sky}: Sky condition
\item \strong{precip}: Precipitation
\item \strong{cost_repairs_infl_adj}: Cost of repairs adjusted for inflation
}
}
\source{
\href{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2019/2019-07-23/wildlife_impacts.csv}{tidytuesday GitHub}
}
\usage{
wildlife_strikes
}
\description{
Provides the Dataset for reported wildlife strikes from 1991 to 2018 in the U.S, reported by The New York Times.
}
\keyword{datasets}
